chore: Fix load and reps properties in workoutModel
The code changes in the workoutModel.js file fix the incorrect assignment of the load and reps properties. The load property was mistakenly assigned to reps and vice versa. This commit corrects the assignment to ensure the proper mapping of the properties.

Note: The recent user commits and recent repository commits do not provide any additional information for generating the commit message.