chore: Add requireAuth middleware for authentication
The code changes add a new middleware called requireAuth in the backend/middleware/requireAuth.js file. This middleware is responsible for verifying the authentication of incoming requests. It checks for the presence of the "authorization" header and extracts the JWT token. If the token is valid, it sets the "req.user" property with the user ID. Otherwise, it returns a 401 error response.

Note: The recent user commits and recent repository commits do not provide any additional information for generating the commit message.