chore: Update CORS origin handling in server.js
The code changes in the server.js file update the CORS origin handling. Instead of using a single origin specified in the environment variable `CORS_ORIGIN`, the code now allows multiple origins specified in the `allowedOrigins` array. If the request origin is in the `allowedOrigins` array or if there is no origin specified, the request is allowed. Otherwise, a CORS error is returned.

Note: The recent user commits and recent repository commits do not provide any additional information for generating the commit message.